Create Base Facet
Still need to migrate basic implementation for Quantities and Units
