+++ pint/facets/base/objects.py
@@ -1,0 +1,13 @@
+# TODO: add basic implementation
+class Unit:
+    pass
+
+
+# TODO: add basic implementation
+class Quantity:
+    pass
+
+
+# TODO: Remove this from here
+class Context:
+    pass
