+++ pint/facets/context/registry.py
@@ -15,7 +15,7 @@ from typing import Any, Callable, ContextManager, Dict, Union
 
 from ..._typing import F
 from ...errors import DefinitionSyntaxError, UndefinedUnitError
-from ...facets.base.definitions import UnitDefinition
+from ...facets.base import UnitDefinition
 from ...util import find_connected_nodes, find_shortest_path, logger
 from .definitions import ContextDefinition
 from .objects import Context, ContextChain
