+++ pint/testsuite/test_converters.py
@@ -2,7 +2,7 @@ import itertools
 
 from pint.compat import np
 from pint.converters import Converter
-from pint.facets.base.definitions import ScaleConverter
+from pint.facets.base import ScaleConverter
 from pint.facets.nonmultiplicative.definitions import (
     LogarithmicConverter,
     OffsetConverter,
